refactor: refactor auth for ticket integrations
